requirements.txt
gymnasium
stable-baselines3
python-chess
pygame
PyOpenGL
numpy
matplotlib